moving spaces and tabs

+++ src/enzo/EvolvePhotons.C
@@ -727,17 +727,21 @@ int EvolvePhotons(TopGridData *MetaData, LevelHierarchyEntry *LevelArray[],
     /* Clean up photon sources if time is up */
     while (RS != NULL){
       if ( ((RS->CreationTime + RS->LifeTime) < PhotonTime) && LoopTime == TRUE &&
-	           DeleteSources) {
-      if (debug) {
-         fprintf(stdout, "\nEvolvePhotons: Deleted Source on lifetime limit \n");
-	       fprintf(stdout, "EvolvePhotons:  %"GSYM" %"GSYM" %"GSYM" \n",
-		             RS->CreationTime, RS->LifeTime, PhotonTime);
-      }
-	       RS = DeleteRadiationSource(RS);
-         NumberOfSources--;
+             DeleteSources) {
+
+        if (debug) {
+          fprintf(stdout, "\nEvolvePhotons: Deleted Source on lifetime limit \n");
+          fprintf(stdout, "EvolvePhotons:  %"GSYM" %"GSYM" %"GSYM" \n",
+                                             RS->CreationTime, RS->LifeTime, PhotonTime);
+        }
+
+        RS = DeleteRadiationSource(RS);
+        NumberOfSources--;
+
       }  else {
         RS = RS->NextSource;
       }
+
     } // end while RS loop
   } // end source clean up
